m1-init: fix all points 1) наличие корневого проекта - в master; 2) наличие подпроектов на Kotlin - m1; 3) правильно настроенный .gitignor и отсутствие мусорных файлов в проекте - m1; 4) версии библиотек в gradle.properties - done in m1.
