+++ gradle.properties
@@ -1,2 +1,3 @@
 kotlin.code.style=official
+jupiterVersion=5.8.1
 kotlinJvmVersion=1.7.10
