+++ m1/build.gradle.kts
@@ -1,17 +1,5 @@
-plugins {
-    id("java")
-}
+group = "ru.otus.homework1"
 
-group = "ru.otus"
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation("org.junit.jupiter:junit-jupiter-api:5.8.1")
-    testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.8.1")
-}
 
 tasks.getByName<Test>("test") {
     useJUnitPlatform()
