+++ m1/src/main/kotlin/ru/otus/homework1/Main.kt
@@ -1,0 +1,9 @@
+package ru.otus.homework1
+
+fun main () {
+    println("Hello, World!")
+}
+
+fun sum (a : Int, b : Int) : Int {
+    return a + b
+}
