--- m1/src/main/kotlin/ru/otus/otuskotlin/Main.kt
@@ -1,5 +1,0 @@
-package ru.otus.otuskotlin
-
-fun main () {
-    println("Hello, World!")
-}
