+++ m1/src/test/kotlin/ru/otus/homework1/MainKtTest.kt
@@ -1,0 +1,12 @@
+package ru.otus.homework1
+
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
+
+internal class MainKtTest {
+
+    @Test
+    fun shouldSumReturn10From5And5() {
+        assertEquals(/* expected = */ 10, /* actual = */ sum(a = 5, b = 5))
+    }
+}
